fix: switch to docker-outside-of-docker
There are a number of packages (eg, homebrew and nix) that don't
support the `linux/arm64` platform. This feels like a warning against
using that platform. The only thing I found that didn't work when running
`linux/amd64` on my Mac was the docker-in-docker package, so I've
switched to using the docker-outside-docker package to use the docker
command on the host machine.

This is a dependency for devcontainers anyway, so feels like a low-risk
move.